Fix issues found by static analysis (#183)
* Fix bad integer size for indexes, which are able to cause overflow

* Move ini implementation to single file, for easy exclusion in semmle

* Don't open files without specifying a umask

* Fix some issues with NULL deref errors

* Fix NULL check for version

* Remove useless GetAutoFailoverNode

* Make sure pid does not overflow

* Make sure pgport does not overflow

* Handle malloc returning NULL

+++ src/monitor/node_metadata.c
@@ -63,7 +63,7 @@ AllAutoFailoverNodes(char *formationId)
 	};
 	const int argCount = sizeof(argValues) / sizeof(argValues[0]);
 	int spiStatus = 0;
-	uint32 rowNumber = 0;
+	uint64 rowNumber = 0;
 
 	const char *selectQuery =
 		"SELECT * FROM " AUTO_FAILOVER_NODE_TABLE " WHERE formationid = $1";
@@ -190,7 +190,7 @@ AutoFailoverNodeGroup(char *formationId, int groupId)
 	};
 	const int argCount = sizeof(argValues) / sizeof(argValues[0]);
 	int spiStatus = 0;
-	uint32 rowNumber = 0;
+	uint64 rowNumber = 0;
 
 	const char *selectQuery =
 		"SELECT * FROM " AUTO_FAILOVER_NODE_TABLE
